python/ec2_inventory/requirements.txt
boto3==1.24.81
botocore==1.27.81
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
